requirements.txt
Django>=3.2.3,<3.3
psycopg2>=2.8.6,<2.9
djangorestframework==3.12.2
django-cors-headers==3.5.0
dj-database-url==0.5.0
python-dotenv==0.15.0
djangorestframework_simplejwt>=5.0
